Organize into submodules and add prod method
Because not all of these methods seem to fit together. For example, some are
seeming more 'mathy' and others more 'testing-y'. I expect that I'll be adding
many more methods and tools to this kit, so in preparation, I'm adding some
organization.

Also adds the prod method to the math module.